Add vocabulary concept

Add a new concept in the database concept table

Usage

add_vocabulary_concept(
  output,
  m = shiny::reactiveValues(),
  vocabulary_id = character(),
  concept_name = character(),
  domain_id = character(),
  concept_class_id = character(),
  concept_code = character(),
  i18n = character(),
  ns = character()
)

Arguments

output

Shiny output variable

m

A shiny::reactiveValues object, used to communicate between modules

vocabulary_id

Value of vocabulary_id where the concept will be added (character)

concept_name

Value of concept_name (character)

domain_id

Value of domain_id, must be present in database (character)

concept_class_id

Value of concept_class_id, must be present in database (character)

concept_code

Value of concept_code (character)

i18n

Translator object from shiny.i18n library

ns

Shiny namespace

Details

See OMOP concept table for more details.

Vocabulary_id must be present in the database when you add a new concept.
Not null values are required for concept_name, domain_id, concept_class_id and concept_code.
The value you provide for domain_id & concept_class_id must be present in the database.

Examples

if (FALSE) {
add_vocabulary_concept(output = output, m = m, i18n = i18n, ns = ns, vocabulary_id = "MIMIC-IV", 
  concept_name = "Diuresis", domain_id = "Measurement", concept_class_id = "Clinical Observation", concept_code = "Diuresis")
}