Import a dataset

Import a dataset in the application, with OMOP CDM

Load +/- save data when a dataset is selected by the user.

Usage

import_dataset(
  output,
  ns = character(),
  i18n = character(),
  r = shiny::reactiveValues(),
  d = shiny::reactiveValues(),
  dataset_id = integer(),
  data = tibble::tibble(),
  type = "",
  omop_version = "6.0",
  save_as_csv = TRUE,
  rewrite = FALSE,
  quiet = TRUE
)

Arguments

output

variable from Shiny, used to render messages on the message bar

ns

Shiny namespace

i18n

shiny.i18n object for translations

r

A shiny::reactiveValues object, used to communicate between modules

d

A shiny::reactiveValues object, used to communicate between modules

dataset_id

ID of the dataset, used to create directory in data folder (integer)

data

Data variable (data.frame or tibble)

type

Name of the OMOP table to import (character)

omop_version

OMOP versions of the imported data, accepts "5.3", "5.4" or "6.0" (character)

save_as_csv

Save or not the data to CSV file (logical)

rewrite

If save_as_csv is TRUE, rewrite or not existing CSV file (logical)

quiet

Should message bars be displayed (logical)

Details

The function is used in a dataset code and is launched each time a user selects a dataset.

You can choose to load data each time the function is used with save_as_csv set to FALSE (eg when dataset is small and the data is loaded from a data warehouse) or you can save the data in a CSV file with save_as_csv set to TRUE.

The data you import must respect the format of OMOP common data model.

See help pages in the app for more informations.

Examples

if (FALSE) {
persons <- tibble::tibble(person_id = 1L, gender_concept_id = 8532L, year_of_birth = NA_integer_, month_of_birth = NA_integer_,
  day_of_birth = NA_integer_, birth_datetime = lubridate::ymd_hms("1990-01-03 00:00:00"), race_concept_id = NA_integer_,
  ethnicity_concept_id = NA_integer_, location_id = NA_integer_, provider_id = NA_integer_, care_site_id = NA_integer_,
  person_source_value = NA_character_, gender_source_value = "F", race_source_value = NA_character_, ethnicity_source_value = NA_character_)
    
import_dataset(output = output, ns = ns, i18n = i18n, r = r, d = d, dataset_id = 5, data = persons, type = "persons", 
  omop_versions = "5.4", save_as_csv = FALSE, rewrite = FALSE, quiet = TRUE)
}