Save updates in datatable

Save updates in a datatable in settings pages

Usage

save_settings_datatable_updates(
  output,
  r = shiny::reactiveValues(),
  m = shiny::reactiveValues(),
  ns = character(),
  table = character(),
  r_table = character(),
  duplicates_allowed = FALSE,
  i18n = character(),
  r_message_bar = FALSE
)

Arguments

output

Shiny output variable

r

Shiny reactive value, used to communicate between modules (reactiveValue)

m

Shiny reactive value, used to communicate between modules (reactiveValue)

ns

Shiny namespace

table

Name of the database table (character)

r_table

Name of the r variable containing the data (character)

duplicates_allowed

Are duplicates in the name column allowed (logical)

i18n

Translator object from shiny.i18n library

r_message_bar

If the MessageBar is displayed with a r variable or with show_message_bar function (TRUE for r variable)

Examples

if (FALSE) {
save_settings_datatable_updates(output = output, r = r, ns = ns, table = "datasets", 
  r_table = "datasets", duplicates_allowed = FALSE, i18n = i18n, r_message_bar = FALSE)
}