Show or hide cards

Show or hide cards depending of selected shiny.fluent::PivotItem

Usage

show_or_hide_cards(
  r = shiny::reactiveValues(),
  session,
  input,
  table = character(),
  id = character(),
  cards = character()
)

Arguments

r

Shiny reactive value, used to communicate between modules (reactiveValue)

session

Shiny session variable

input

Shiny input variable

table

Name of the database table (character)

id

Name of the ID of the page (character)

cards

Character vector containing names of current page cards (character)

Examples

if (FALSE) {
cards <- c("subsets_datatable_card", "subsets_persons_card", "subsets_edit_code_card")
show_or_hide_cards(r = r, input = input, session = session, id = id, cards = cards)
}